agent-tools/sec-edgar/requirements.txt
# Dependencies specific to the SEC Edgar tool
boto3
edgartools==4.0.1
pytest
requests
langchain-core
langchain-openai
python-dotenv
